Initial implementation of code coverage for 1.19.2 tests
We have to make everything use the mojang-mapped classifier. In time, we'll refactor this to have the relevant modules export this classifier, so our testing environments can use it. We'll also have to find a way to switch our reflection mappings to support this as well. It's obviously important to have some tests for ensuring that the non-reflected mappings are valid as well, which we can probably load as an additional profile with a non-mapped instance (of say commandapi-shade)